+ uses throw/catch instead of raise/rescue
This is done for speed reasons: Stack traces cost a lot. This version is
about an order of magnitude faster than the last one. No change for the
user, he still only sees a ParseFailed exception, but always with the
same stack trace. (which is good, since that stack trace is internal
anyway)

the computer
you've used for years
suddenly so slow