+ prsnt? -> present?, absnt? -> absent?
Thanks to txus (Josep M. Bach) for nudging me ever so slightly into that
direction and for providing a patch. I rewrote it to have it exactly my
way, which is not saying that the patch was bad. It just testifies my
perfectionism. Thanks a lot, it makes sense to have it this way.