Modernizes dependencies and rspec code
The blankslate dependency is now more flexible to integrate with other
projects.

RSpec-code has been given a rehaul to eliminate all deprecation
warnings on modern rspecs. I must have broken compat to older versions
in the process, but that complaint should be addressed to the rspec
project, not me. (really, guys!)