Document infix_expression on the website
I requested this in https://github.com/kschiess/parslet/issues/135 after
scouring the parser and tricks pages of the website looking for some
assistance in implementing operator precedence with parslet. I know
there's an example in the examples, and it's documented within the
codebase, but just from my experience, I felt like if it wasn't on the
website, it didn't exist. Once I found it (via the Salama project) I was
like, weeping with gratitude.