Help the rspec rig run under RSpec 2.99
I'm currently upgrading my app from RSpec 2.14 to the latest. [The upgrade
guide](https://relishapp.com/rspec/docs/upgrade) recommends first upgrading to
2.99, then resolving all deprecation warnings, and then upgrading to 3.

Strangely, the rspec rig works great on 2.14 and 3.3, but fails on 2.99 with:

```
NameError:
  undefined method `failure_message_for_should' for class `Class'
```

* here's a small project demonstrating the problem with the latest
  parslet and RSpec 2.99: https://github.com/maxjacobson/parslet_failure_demo
* here's a branch demonstrating that this patch works on RSpec 2.14
  https://github.com/maxjacobson/parslet_failure_demo/tree/rspec-2.14
* here's a branch demonstrating the patch works with RSpec 2.99
  https://github.com/maxjacobson/parslet_failure_demo/tree/rspec-2.99
* here's a branch demonstrating the patch works with RSpec 3.3
  https://github.com/maxjacobson/parslet_failure_demo/tree/rspec-3.3

This commit adds an explicit rspec version expectation for testing the
parslet source code. That should be safe to fix in place, but parslet
cannot know what version of rspec is used by the parslet-wielding
developer and is gracious to support multiple.