Removed `abort` and `catch_error` methods from infix atom
These methods were just wrappers around `throw` and `catch` that
obscured the logic.  The `abort` call referenced in the orignal issue
was really a call to the unfortunately named helper and not to
`Kernel.abort`.

I have removed the helpers, which should make the code more explicit,
without impacting functionality.

Fixes #186